Install now works completely.
If config.h is included in lcsio.cpp, it will append PREFIX/share/lcs/art to the
paths to search for art files so it can find it is odd locations.

configure.ac was modified to define INSTALL_PREFIX in config.h

Makefile.am was modified to properly install the art files.