The "fixcleartype" option in init.txt has been made much safer.
Old settings are backed up to disk, and settings are only changed if the backup is successful. If the game crashes, old settings are read from disk. Old settings are now restored whenever the game ends with a call to end_game() from anywhere in the game, not just when the call to end_game() that ends the game is the one in main() in game.cpp. After restoring old settings successfully, the backup of those settings to disk is deleted.