Fixed bug pointed out by SlatersQuest on the forum.
The difficulty to find a mutant with the recruitment activity now corresponds better with the chance of finding one in sitemode. Easy when nuclear and pollution laws are both C+, "Hard" when only one of them are C+ and "Nearly impossible" when neither are C+.

Though you are still guaranteed to always find at least one. (daily/recruit.cpp, line 125)