More names added to all the name lists (ethnic names, celebrity names, names of serial killers, names of political figures, and various miscellaneous names).
The very very short last names (like the shortest of all, "O") are all extremely common Chinese or Korean names, most of which seem to be 2 or 3 letters. And the names with accents, those are foreign, obviously... the names are more ethnically diverse now, from all over the world. Except for white male patriarch names of course, they are not diverse at all.