Improved monthly funding report:
Amounts are all right-justified so the digits line up.
Daily change is shown as well, along the right side.
'$' key can be used to display the funding report any time, not just at the end of the month.

Added a categorized asset tally, describing the amount of money which can be quickly raised at the pawn shop.