New file with LCS lyrics to national anthem: /dev/national_anthem_lyrics.txt.
The point of this is, a victory song for the end of the game, to the tune of the U.S. national anthem, to replace victory.ogg. Now all we need is someone to record a performance of the national anthem with these lyrics, as a lossless 44100 Hz FLAC audio file that they release to the public domain and contribute to this project, and I'll take care of making an Ogg Vorbis version that has any necessary filters applied to it to make it sound as good as possible (noise removal, dynamic range compression, etc.), without altering the original FLAC recording.