Refactored once again. Can now start augmentations
Originally, all the augmentations were stored in enums within the creature class. That wasn't the best idea, so now I moved the augmentations to an xml file within the art/ directory.
There are four new classes (two .h and two .cpp within src/creature/): one pair for extracting the data from xml and the other for the actual augmentation class, so it's consistent with the project in general.