Don't clobber user variables in autotools builds
The variable $CSSFLAGS is reserved for user use and should not be forced by the
build machinery (like the autotools).  This prevents using fancy-schamncy things
like -Wall from the command line.

This change moves a few things around so that the flags required by the build
machinery (eg. -std=c++11) use the AM_CXXFLAGS reserved for that purpose and
leaves CXXFLAGS free for appropriate use.