Add separate autoconf check for langinfo.h
The includsion of the POSIX header <langinfo.h> was dependent on the use of
autoconf (via HAVE_CONFIG_H) rather than through the presence of the header
itself.

Tsk, tsk.  Causes a FTBFS in a mingw64 cross compile.

Added an explicitl check on the header to configure.ac and used the result to
option out the includion of <langinfo.h> from the nightmarish "common.h".