Remove AC_FUNC_MALLOC because it breaks C++
The AC_FUNC_MALLOC directive is used to detect libc-compliant functionality of
malloc() (which is implementation-defined by the C standard.  Unfortunately, it
assumes that if you;re cross-compiling, you don;t have such an implementation,
and unfortunately that just breaks libstdc++ (the C++ standard library that
comes with GCC).

The choice is to either work around this situation with macro hackery and
replacement implementations, or just remove the directive from configure.ac.  A
wuick inspection of the code shows that it does not rely on libc-dependent
behaviout anywhere (at least that I can find), so this change takes the second
approach.