Don't redefine some ncurses symbols
Earlier versions of ncurses may not have handled UTF-8 characters correctly or
something (the comments are unrevealing) and some functions were actually
defined macros instead of real functions.  As of ncurses 6.0, UTF-8 handling is
pretty good and those self-same functions are real functions, not macros, so
it's wrong to try to violate the ODF.

The problem of course lies in the bletcheriffic "includes.h" megaheader.