Fix -Wunused-result warnings
The injudicious use of fread without checking its return value was filling the
build logs with unsolicited noise.  While it would be better to actually check
the result of these I/O operations, in the meantime the warnings can be silenced
through the wicked and scurrilous use of pragmas.

Also, moved some news functions out of includes.h and into the news.h header.